+++ src/containers/Sidebar/SidebarBottomItem.js
@@ -14,7 +14,7 @@ import Icon from './SidebarIcon';
 
 import { SET_SMALL_SIDEBAR } from '../../actions';
 
-const SidebarBottomItem = ({ item, sidebarLogo }) => {
+const SidebarBottomItem = ({ item, sidebarLogo, screenWidth }) => {
 	const isSidebarCollapsed = useSelector(state => state.sidebar.isSidebarCollapsed);
 	const isSmallSidebarOpen = useSelector(state => state.sidebar.isSmallSidebarOpen);
 
@@ -42,6 +42,13 @@ const SidebarBottomItem = ({ item, sidebarLogo }) => {
 		});
 	}
 
+	if (screenWidth && screenWidth <= 768) {
+		dispatch({
+			type: COLLAPSE_SIDEBAR,
+			isSidebarCollapsed: false
+		});
+	}
+
 	const AboutButton = () => {
 		return (
 			<>
@@ -59,7 +66,7 @@ const SidebarBottomItem = ({ item, sidebarLogo }) => {
 						{item && (
 							<button
 								title={t(`Sidebar|${item.name}`)}
-								className={`sidebar-item-button${location.pathname === item.url ? " active " : " "}btn left mobile-full-btn`}
+								className={`sidebar-item-button${location.pathname === item.url ? " active " : " "}btn left ${!isSidebarCollapsed && "mobile-full-btn"}`}
 							>
 								{sidebarLogo ? 
 									isSidebarCollapsed ?