+++ src/containers/Sidebar/index.js
@@ -63,16 +63,16 @@ const Sidebar = (props) => {
 	const toggle = () => setModal(!modal);
 
 	return (
-		windowDimensions.width < 768 ?
+		windowDimensions.width <= 768 ?
 		<>
 			<div className="mobile-sidebar-burger" onClick={toggle}>
 				<i className="cil-menu"></i>
 			</div>
 			<Modal isOpen={modal} toggle={toggle} className="left">
-				<div className={`app-sidebar${isSidebarCollapsed ? " collapsed" : ""}`}>
-					{windowDimensions.width > 768 &&
+				<div className="app-sidebar">
+					{windowDimensions.width >= 768 &&
 						<div style={{display: "inline-block"}}>
-							<NavbarBrand {...props} isSidebarMinimized/>
+							<NavbarBrand {...props}/>
 						</div>
 					}
 					<div className="sidebar-nav">
@@ -87,7 +87,7 @@ const Sidebar = (props) => {
 								/>
 							))}
 						</Nav>
-						<SidebarBottomItem item={aboutItem} sidebarLogo={props.sidebarLogo} />
+						<SidebarBottomItem item={aboutItem} sidebarLogo={props.sidebarLogo} screenWidth={windowDimensions.width}/>
 					</div>
 				</div>
 			</Modal>